# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Workshop OS API",
        "description": "API para gerenciamento de ordens de serviço de oficina mecânica. Perfis: receptionist, mechanic, client.",
        "contact": {
            "email": "dev@workshop.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/v1/service-orders": {
            "get": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Lista ordens de serviço",
                "operationId": "37002ce155bedc30b00018237d47d174",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de OS"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Cria uma Ordem de Serviço",
                "operationId": "7f17c440e6edc12a5fac1bdb151898a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "vehicle_id"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OS criada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}": {
            "get": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Exibe OS com serviços e peças",
                "operationId": "333f09eaf953dc171b140ac638e192ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados completos da OS"
                    },
                    "404": {
                        "description": "OS não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/services": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Adiciona serviço à OS (mecânico)",
                "operationId": "e528985859357de8c27c7686d5016e52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serviço adicionado"
                    },
                    "422": {
                        "description": "OS não permite alteração neste status"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/parts": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Adiciona peça à OS (mecânico)",
                "operationId": "ea7e6b9cd76568f6eb40f33896b06bb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "part_id"
                                ],
                                "properties": {
                                    "part_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Peça adicionada"
                    },
                    "422": {
                        "description": "Estoque insuficiente ou status inválido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/generate-budget": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Gera orçamento e avança para Aguardando aprovação (mecânico)",
                "operationId": "c8772ba3ee3697bd329a27752159cf44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orçamento gerado e enviado ao cliente"
                    },
                    "422": {
                        "description": "Status inválido para gerar orçamento"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/approve": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Cliente aprova o orçamento",
                "operationId": "951eebb343092171d4a0a42a6d88a08a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OS aprovada"
                    },
                    "422": {
                        "description": "Status inválido para aprovação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/cancel": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Cliente cancela o orçamento",
                "operationId": "ac7390bd19a654c04c57bc734e6625f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OS cancelada"
                    },
                    "422": {
                        "description": "Status inválido para cancelamento"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/start-execution": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Inicia a execução (mecânico)",
                "operationId": "096a8718e486099dc230d6176cbf8829",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OS em execução"
                    },
                    "422": {
                        "description": "Status inválido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/finalize": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Finaliza execução (mecânico)",
                "operationId": "74b538e472cc951f80d9e3831b6559c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OS finalizada — cliente notificado para retirar"
                    },
                    "422": {
                        "description": "Status inválido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/service-orders/{id}/pay": {
            "post": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Cliente paga a OS",
                "operationId": "7d98a7d92cd2cdd704e2f0a17f21bb86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento processado"
                    },
                    "422": {
                        "description": "OS não está Finalizada ou já foi paga"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/service-orders/stats": {
            "get": {
                "tags": [
                    "ServiceOrders"
                ],
                "summary": "Monitoramento: tempo médio de execução por serviço",
                "operationId": "80a7c0d1b76e2bb95fb580e5c831c560",
                "responses": {
                    "200": {
                        "description": "Estatísticas de execução"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Informe o token Sanctum: Bearer {token}"
            }
        }
    },
...
}